Add cover-only and replace pattern handling of main function.
This commit adds a new command-line option in the code (not on the
command line itself), `cover-only`, which takes either `function` or
`file` as parameter. The former is equivalent to `cover-function-only`,
the later instruments the file which contains the main function.

As part of the changes, instead of matching the function name using
regex for coverage instrumentation we use the actual function symbol.