Refactor `java_build_arguments` to return a `pair`
`java_build_arguments` was previously returning the argument expressions
as the return value and the code which created these by out parameter.
This refactor changes it to return both as the return value. This makes
it clearer that the implementation does not depend on any previously
generated code and clear where output goes to.