Make failed nondet Enum assignment non-fatal
Previously we always expected an Enum type's VALUES array to be present if it may be created in a
stub method. However while this is easy for a front-end to ensure when the Enum is a direct return
type or member of a stub return type, there are other cases where the object factory can end up
building a nondet Enum, such as generic type substitution and usage-inferred stub return types.
Until (if ever) ci-lazy-methods can precisely mimic the object factory's behaviour, we're better
off tolerating the case where the VALUES array remains missing than killing the entire run.