Remove redundant code in ID_dynamic_object branch
This branch is not taken by symex but it is taken by the slicer. The code
I've removed has no effect, as far as I can tell, and may have been left
over from an old implementation from before the beginning of the git
history.

Note, there is a commit that splits ID_dynamic_object into two different
ids. See #2646 and #3769 for more details.