Read "is inlined" attribute of goto functions from the symbol table
There should only be a single place to hold type information, including
attributes, to ensure consistency. Future changes will remove the "type" member
of goto_functiont, making the type information stored in the symbol table the
single, authoritative source of information.

The "is inlined" information should already be consistent/redundant. This commit
makes all read accesses use the information stored in the symbol table, and also
uses a modern API for doing so.