Split out goto_statet::source
This is the one member of goto_statet which is not trivially moved when it is combined
with an existing goto_symex_statet -- rather it labels the incoming state and is discarded
after perhaps being inspected by the merge_goto routine. By splitting these we prepare
goto_statet to use a trivial move-constructor when merging into a goto_symex_statet with
a false state guard.

There should be no behavioural change caused by this commit.