Refactor sharing map
This changes the interface of the sharing map to make it easier to use. Now the
only method where sharing might be broken unnecessarily is replace(). A debug
mode is also provided to detect such cases. As a tradeoff, the new interface now
further deviates from the interface of std::map.