Remove unnecessary location field from object factory
An instance of java_object_factoryt is only constructed in two places in
the code. Both times, gen_nondet_init is immediately called on it with
the same source location that was also passed to the constructor.
Storing the same value both in a field and in function parameters is
unnecessary, so the field is removed in this commit.