Remove unnecessary namespacet field from object factory
ns was supposed to always contain a pointer to symbol_table. This kind
of class invariant is unnecessary. It is better to just work with the
symbol table if possible, and construct a namespacet from it whenever we
need it. This way the namespacet can always be declared const and we can
easily see which functions make use of it.