Remove the find_state internal interface from ai_baset
find_state is not currently used nor is it likely to be.
During computation, *this is non-const so get_state can be used.
Output methods now use the public interface of
abstract_state_before.
The only real space for it to be used is in ait::operator[] const
which can access the map directly so doesn't need to have an
additional interface in ai_baset.