Use owning class functions as interface to the owner of a symbol
This commit replaces access to `symbol.type.get(ID_C_class)`. With calls
to the `owning_class` function and access to
`symbol.type.set(ID_C_class, class_id)` with calls to
`set_owning_class`. This allows for changes to how the owner of a symbol
is stored by changing these two functions only, rather than by changing
all the places where `ID_C_class` was previously accessed directly. This
is preparatory work for moving this link from being stored in a comment
on a symbols type, to the symbol itself.