Set declaring class for string solver implemented methods
The declaring class was not previously set for methods which are
implemented by the string solver. This commit sets the declaring class
for the initially empty valued method symbols. INVARIANTs are used to
check that the declaring class is still set once the value of the symbol
has been provided for the methods implemented by the string solver. The
non violation of this INVARIANT is tested by all `strings-smoke-tests`
which call such methods. For example the tests in
`regression/strings-smoke-tests/java_int_to_string` will check that the
INVARIANT holds for `Integer.toString`.