Create a domain factory object so domains can have constructors
The lack of a way of passing configuration options and the like
to domains from outside of the abstract interpreter has been a
long-standing pain.  This moves domain construction into a factory
so that users of the abstract interpreter can just inherit from
ai_domain_factory_baset or ai_domain_factoryt<domainT> to have
an object that carries the extra constructor arguments and then
uses them when needed.