Refactor add_axioms_for_format_specifier
This makes it accept a function as argument, which gives for possible
types the corresponding value of the argument.
This will allow us to adapt more easily to a change in the way the
arguments of format are given: they are now given as a struct but we
want to have the possibility to give them as a string.