C++ front-end: put "this" parameter in the symbol table
The "this" parameter must carry a unique name for each class and must be
stored in the symbol table. And indeed make sure that all parameters are
in the symbol table, unless the method a member of a class template and
is never used, in which case the method symbol must be removed.
Previously we mostly did this removal, but would fail for methods with a
body not involving any symbols (as we tested for the existence of some
cpp_name).