Fix strings not being recognized in the object factory
For a given struct, the object factory previously checked if it was a
string only in the case where skip_classid was set to false and the
update strategy was NO_UPDATE_IN_PLACE or MAY_UPDATE_IN_PLACE.
This missed some cases, for example the case where no model for String
is loaded and the Java method includes a call to a string constructor
that depends on a model. In this case the constructor is stubbed and
skip_classid is set to true in java_simple_method_stubst. Before this
fix, the object factory would go through the components of the string
expression as if it was a regular object expression, and hit an
invariant violation because the data component is stored as a pointer
to an unsignedbv, contradicting the assumption that all pointers point
to structs.