boolbvt: introduce scopes
This is the biggest change to the solver interface since it got added to the
codebase.

Currently, bound identifiers (exists, forall, let, lambda,
array_comprehension_exprt) must be unique; the solver loses soundness if
that assumption is violated.

This change introduces scopes for let, i.e., the uniqueness requirement is
dropped.  Nested scoping works as expected.  This matches what SMT-LIB does,
i.e., the renaming done in the SMT-LIB front-end can be dropped.