Export-function-local-symbols name mangling: rename initial values
We previously renamed all occurrences in goto functions, but did not
consider the initial values of symbols with static lifetime, which are
stored in the symbol table. To make this reasonably efficient, make
rename_symbolt::operator() expose the return value (and document the
methods).

To test this, include the test suite in Makefile-based test set-ups
(CMake was already done), and fix the chain.sh file.