Symex: implement lift_lets
This executes let expressions like ordinary assignments, avoiding the need for all other components
of cbmc to understand the little-used local definition expression. It also means that symex's usual
constant propagator, value-set and other analyses are brought to bear on the definition just as if
the front-end had used an instruction to make the definition.