Add non-negative length constraint earlier
Add non-negativeness constraint on string lengths before doing the first
iteration over the index set.

These constraints used to be added after that step, and some of the
length variables were missing these constraints.

This is made very visible by the upcoming changes to the way we deal
with string lengths.