Add special case for deref-of-typecast-of-if
This results from expressions such as *(int*)*p, and is very nearly as easy to handle as the
already-special-cased deref-of-if case. In essence we transform *(type*)(x ? y : z) into
x ? *(type*)y : *(type*)z, which should yield a more compact expression due to not conflating
the alias sets that can be derived from the two arms of the conditional.