Add attempt_assign_length_from_type
Function to add the length of the string backing array to the
length_of_array map in the array_pool. If the size does not exist,
create a symbol.

Function will be used to assign an initial length to the
array_string_exprt, using the type information.

After that initial assignment, there will be no further reference to the
length of the underlying array.