Reorder types in sharing_nodet to optimize for common case
When destructt::destruct() of a small shared n-way pointer is called, it first
determines the type of its pointee. For example, if the parameter pack Ts...
is instantiated to T1, T2, T3, then it first checks if the pointer points to T3,
then to T2, etc. Thus, the least checks are performed if it is most likely that
the pointer points to an object of type T3, then T2, etc.