Rename lambda_exprt to array_comprehension_exprt
This kind of exprt does not represent lambda functions, but it
represents an array by specifying a range and a function operating on
the array indices.
This is a type of list/array comprehension, so calling it
array_comprehension_exprt is more descriptive of what this type is used
for.