User-provided function definitions take precedence over built-ins
We should not replace calls to functions that have a body provided by
the user by our built-in definitions. While function names starting with
"_" are reserved in C and thus users shouldn't be using them, it is
certainly not true that anything named "nondet_*" is exclusively to be
captured by our built-ins.