introduce simplify_exprt::resultt
This is follow-up on #2939 and #2946, and the biggest change to the
simplifier since its existence.

It attempts to make the implementation of individual simplification steps
more type safe, as opposed to using exprt.

The alternative considered was to return optionalt<exprt>, which has the
disadvantage that it doesn't "chain", i.e., given two transformations f and
g one cannot write f(g(...)).