Check pointer evaluation in the member case
Field sensitivity the value set to keep track of pointers in individal
fields, making it possible to evaluate equalities such as the one in the
example.
This is a case which would not work if try_evaluate_pointer_comparisons
was using get_l1_object_identifier instead of getting the identifier of
the whole l1 expression (not the underlying object).

fixup! Check pointer evaluation in the array case