Add --dump-c-type-header to goto-instrument
The --dump-c-type-header flag is going to dump a header file
containing the file local type definitions in a c file.
While it might be desirable to keep a type definition
only in a c module, for some proofs it is required to
include the type definition supporting the harness setup.
If the file local type definition requires to include types defined
in other header files, an include for the other header file
is added as well.