Add `code_outputt` class to wrap around `ID_output` based ireps
This commit adds a higher level interface for `ID_output` based ireps.
This gives us a central place to document instances of these and a
central place to put the code for constructing and checking them. This
makes it possible to find documentation about them and avoids
duplicating the code for constructing and checking them.