Do not use backtick as opening single quote
Replace it with '. For example, This changes an error message from
failed to load class `java.nio.file.Path'
to
failed to load class 'java.nio.file.Path'

The reason is that the convention of using ` for opening single
quotes looks odd in most modern fonts. For more discussion please
read:

https://www.cl.cam.ac.uk/~mgk25/ucs/quotes.html