Remove duplicate initialisation during instrumentation
For on-the-fly domain creation to work get_state shouldn't have any
observable side-effects.  At the moment this is not the case because
the operator[] access will throw an exception if the location is not
initialised.  Thus we can't remove both of the calls to get_state
but we can definitely remove these duplicate ones.